Added Controller Support
Left stick moves the player, right bumper is jump, right buttons change colors, a button is used as enter for menu navigation. Right now the left stick movement controls feel a bit off compared to keyboard and also if there is even a tiny bit of analog stick drift the character will move in the direction of the drift.